Switch from inline password to users json
The biggest problem here is environment interpolation.  The caddy
password format contains dollar signs ($) as a separator.
Unfortunately, dollar signs have special meanings in interpolation, and
this is going through several layers of interpolation, making it really
obnoxious to get it escaped.  It was much easier to just yank the
users.json out and map that in instead.